Partial refactoring with Element abstract class
