+++ src/Results/Movie.php
@@ -2,6 +2,7 @@
 
 namespace vfalies\tmdb\Results;
 
+use vfalies\tmdb\Abstracts\Results;
 use vfalies\tmdb\Tmdb;
 
 class Movie extends Results
